+++ src/styles/main/playing/Keyboard.css
