+++ src/styles/main/playing/Timer.css
